requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.0.1
fastapi==0.135.1
fastjsonschema==2.20.0
filelock==3.24.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyMongo==3.0.1
fonttools==4.53.1
fqdn==1.5.1
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.66.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.5
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.13.0
joblib==1.5.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mysql-connector-python==9.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
openai==2.21.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.27.1
protobuf==5.29.6
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pymongo==4.16.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2026.2.19
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.19.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.14.0
selenium==4.28.1
Send2Trash==1.8.3
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.4.1
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
trio==0.28.0
trio-websocket==0.11.1
typer==0.24.1
typer-slim==0.24.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.41.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.3
wsproto==1.2.0
